--- buildSrc/src/test/resources/test_out.md
@@ -1,1 +1,0 @@
-dummy test out file
